feat: adiciona responsividade na tela de listagem de vagas e atualiza imagens do README
