+++ src/components/Card/index.tsx
@@ -64,9 +64,9 @@ export const Card = ({ job }: { job: JobType }) => {
   }, [generateTags]);
 
   return (
-    <CardContainer>
+    <CardContainer tabIndex={0}>
       <Flex gap="24px">
-        <CompanyImage src={job.logo} />
+        <CompanyImage src={job.logo} alt={job.company} />
         <JobDescription>
           <Flex>
             <CompanyName>{job.company}</CompanyName>
