+++ src/components/Header/index.tsx
@@ -1,0 +1,10 @@
+import { ReactComponent as LogoWhite } from 'assets/images/logo-white.svg';
+import { HeaderContainer } from './styled';
+
+export const Header = () => {
+  return (
+    <HeaderContainer>
+      <LogoWhite />
+    </HeaderContainer>
+  );
+};
