+++ src/pages/JobList/styled.ts
@@ -2,12 +2,23 @@ import styled from 'styled-components';
 
 export const Main = styled.main`
   width: 100%;
-  height: calc(100% - 160px);
 
   display: flex;
   flex-direction: column;
   align-items: center;
   gap: 22px;
 
-  padding: 75px 0 30px;
+  margin-top: -30px;
+  @media screen and (max-width: 576px) {
+    gap: 38px;
+  }
+`;
+
+export const AbsoluteBox = styled.div`
+  width: 80%;
+  margin-bottom: 32px;
+
+  @media screen and (max-width: 576px) {
+    margin-bottom: 16px;
+  }
 `;
